+++ randovania/games/prime1/exporter/game_exporter.py
@@ -2,20 +2,52 @@ import copy
 import dataclasses
 import json
 import os
+import shutil
 from pathlib import Path
 
 import py_randomprime
 
 from randovania.dol_patching import assembler
 from randovania.exporter.game_exporter import GameExporter, GameExportParams
+from randovania.game_description.resources.pickup_entry import PickupModel
+from randovania.games.game import RandovaniaGame
 from randovania.lib import status_update_lib
-from randovania.patching.prime import all_prime_dol_patches
+from randovania.patching.patchers.gamecube import iso_packager
+from randovania.patching.prime import all_prime_dol_patches, asset_conversion
+from randovania.games.prime1.exporter.patch_data_factory import _MODEL_MAPPING
 
 
 @dataclasses.dataclass(frozen=True)
 class PrimeGameExportParams(GameExportParams):
     input_path: Path
     output_path: Path
+    echoes_input_path: Path
+    echoes_contents_path: Path
+    echoes_backup_path: Path
+    asset_cache_path: Path
+    use_echoes_models: bool
+    cache_path: Path
+
+
+def adjust_model_names(patch_data: dict, assets_meta: dict, use_external_assets: bool):
+
+    model_list = []
+    if use_external_assets:
+        bad_models = {"prime2_MissileLauncher", "prime2_MissileExpansionPrime1"}
+        model_list = list(set(assets_meta["items"]) - bad_models)
+
+    for level in patch_data["levelData"].values():
+        for room in level["rooms"].values():
+            for pickup in room["pickups"]:
+                model = PickupModel.from_json(pickup.pop("model"))
+                if model.game == RandovaniaGame.METROID_PRIME:
+                    converted_model_name = model.name
+                else:
+                    converted_model_name = "{}_{}".format(model.game.value, model.name)
+                    if converted_model_name not in model_list:
+                        converted_model_name = _MODEL_MAPPING.get((model.game, model.name), "Nothing")
+
+                pickup['model'] = converted_model_name
 
 
 class PrimeGameExporter(GameExporter):
@@ -42,6 +74,9 @@ class PrimeGameExporter(GameExporter):
         input_file = export_params.input_path
         output_file = export_params.output_path
 
+        export_params.cache_path.mkdir(parents=True, exist_ok=True)
+        cache_dir = os.fspath(export_params.cache_path)
+
         symbols = py_randomprime.symbols_for_file(input_file)
 
         new_config = copy.copy(patch_data)
@@ -57,6 +92,22 @@ class PrimeGameExporter(GameExporter):
                 ],
                 symbols=symbols)
         )
+        new_config["preferences"]["cacheDir"] = cache_dir
+
+        assets_meta = {}
+        updaters = [progress_update]
+        if export_params.use_echoes_models:
+            assets_path = export_params.asset_cache_path
+            if asset_conversion.get_asset_cache_version(assets_path) != asset_conversion.ECHOES_MODELS_VERSION:
+                updaters = status_update_lib.split_progress_update(progress_update, 3)
+                from randovania.games.prime2.exporter.game_exporter import extract_and_backup_iso
+                extract_and_backup_iso(export_params.echoes_input_path, export_params.echoes_contents_path,
+                                       export_params.echoes_backup_path, updaters[0])
+            assets_meta = asset_conversion.convert_prime2_pickups(assets_path, updaters[1])
+            new_config["externAssetsDir"] = os.fspath(assets_path)
+
+        # Replace models
+        adjust_model_names(new_config, assets_meta, export_params.use_echoes_models)
 
         patch_as_str = json.dumps(new_config, indent=4, separators=(',', ': '))
         if has_spoiler:
@@ -67,7 +118,7 @@ class PrimeGameExporter(GameExporter):
         try:
             py_randomprime.patch_iso_raw(
                 patch_as_str,
-                py_randomprime.ProgressNotifier(lambda percent, msg: progress_update(msg, percent)),
+                py_randomprime.ProgressNotifier(lambda percent, msg: updaters[-1](msg, percent)),
             )
         except BaseException as e:
             if isinstance(e, Exception):