+++ randovania/server/discord/bot.py
@@ -1,7 +1,10 @@
+import logging
+
 import discord.ext.commands
 from discord_slash import SlashCommand
 
 import randovania
+from randovania.server.discord.randovania_cog import RandovaniaCog
 
 
 class RandovaniaBot(discord.ext.commands.Bot):
@@ -15,6 +18,13 @@ class RandovaniaBot(discord.ext.commands.Bot):
         self.load_extension("randovania.server.discord.preset_lookup")
         self.load_extension("randovania.server.discord.database_command")
         self.load_extension("randovania.server.discord.faq_command")
+
+    async def on_ready(self):
+        for cog in self.cogs.values():
+            if isinstance(cog, RandovaniaCog):
+                await cog.add_commands(self.slash)
+
+        await self.slash.sync_all_commands()
 
 
 def run():