+++ test/games/prime2/gui/test_echoes_game_export_dialog.py
@@ -5,6 +5,7 @@ import pytest
 from PySide6 import QtCore
 
 from randovania.games.game import RandovaniaGame
+from randovania.games.prime1.exporter.options import PrimePerGameOptions
 from randovania.games.prime1.layout.prime_cosmetic_patches import PrimeCosmeticPatches
 from randovania.games.prime2.exporter.game_exporter import EchoesGameExportParams
 from randovania.games.prime2.exporter.options import EchoesPerGameOptions
@@ -32,7 +33,7 @@ def test_on_output_file_button_exists(skip_qtbot, tmp_path, mocker, has_output_d
         output_directory=output_directory,
     )
 
-    window = EchoesGameExportDialog(options, {}, "MyHash", True)
+    window = EchoesGameExportDialog(options, {}, "MyHash", True, [])
     mock_prompt.return_value = tmp_path.joinpath("foo", "game.iso")
 
     # Run
@@ -54,7 +55,7 @@ def test_on_output_file_button_cancel(skip_qtbot, tmpdir, mocker):
         cosmetic_patches=EchoesCosmeticPatches.default(),
         output_directory=None,
     )
-    window = EchoesGameExportDialog(options, {}, "MyHash", True)
+    window = EchoesGameExportDialog(options, {}, "MyHash", True, [])
     mock_prompt.return_value = None
 
     # Run
@@ -65,16 +66,26 @@ def test_on_output_file_button_cancel(skip_qtbot, tmpdir, mocker):
     assert window.output_file_edit.text() == ""
 
 
-def test_save_options(skip_qtbot, tmp_path):
+@pytest.mark.parametrize("is_prime_multi", [False, True])
+def test_save_options(skip_qtbot, tmp_path, is_prime_multi):
     options = Options(tmp_path)
-    window = EchoesGameExportDialog(options, {}, "MyHash", True)
+    games = [RandovaniaGame.METROID_PRIME_ECHOES]
+    if is_prime_multi:
+        games.append(RandovaniaGame.METROID_PRIME)
+    window = EchoesGameExportDialog(options, {}, "MyHash", True, games)
     window.output_file_edit.setText("somewhere/game.iso")
+    if is_prime_multi:
+        skip_qtbot.mouseClick(window.prime_models_check, QtCore.Qt.LeftButton)
+        window.prime_file_edit.setText("somewhere/prime.iso")
 
     # Run
     window.save_options()
 
     # Assert
     assert options.options_for_game(RandovaniaGame.METROID_PRIME_ECHOES).output_directory == Path("somewhere")
+    if is_prime_multi:
+        assert options.options_for_game(RandovaniaGame.METROID_PRIME).input_path == Path("somewhere/prime.iso")
+        assert options.options_for_game(RandovaniaGame.METROID_PRIME_ECHOES).use_external_models == {RandovaniaGame.METROID_PRIME}
 
 
 def test_on_input_file_button(skip_qtbot, tmp_path, mocker):
@@ -102,7 +113,7 @@ def test_on_input_file_button(skip_qtbot, tmp_path, mocker):
         None,
     ])
 
-    window = EchoesGameExportDialog(options, {}, "MyHash", True)
+    window = EchoesGameExportDialog(options, {}, "MyHash", True, [])
     assert window.input_file_edit.text() == "(internal game copy)"
     assert not window.input_file_edit.has_error
 
@@ -136,16 +147,37 @@ def test_on_input_file_button(skip_qtbot, tmp_path, mocker):
     ])
 
 
-def test_get_game_export_params(skip_qtbot, tmp_path):
+@pytest.mark.parametrize("is_prime_multi", [False, True])
+@pytest.mark.parametrize("use_external_models", [False, True])
+def test_get_game_export_params(skip_qtbot, tmp_path, is_prime_multi, use_external_models):
     # Setup
+    games = [RandovaniaGame.METROID_PRIME_ECHOES]
+    if is_prime_multi:
+        games.append(RandovaniaGame.METROID_PRIME)
+        prime_path = tmp_path.joinpath("input/prime.iso")
+    else:
+        prime_path = None
+
+    if use_external_models:
+        models = {RandovaniaGame.METROID_PRIME}
+    else:
+        models = {}
+
     options = MagicMock()
     options.internal_copies_path = tmp_path.joinpath("internal_copies")
-    options.options_for_game.return_value = EchoesPerGameOptions(
-        cosmetic_patches=EchoesCosmeticPatches.default(),
-        input_path=tmp_path.joinpath("input/game.iso"),
-        output_directory=tmp_path.joinpath("output"),
-    )
-    window = EchoesGameExportDialog(options, {}, "MyHash", True)
+    options.options_for_game.side_effect = [
+        EchoesPerGameOptions(
+            cosmetic_patches=EchoesCosmeticPatches.default(),
+            input_path=tmp_path.joinpath("input/game.iso"),
+            output_directory=tmp_path.joinpath("output"),
+            use_external_models=models,
+        ),
+        PrimePerGameOptions(
+            cosmetic_patches=PrimeCosmeticPatches.default(),
+            input_path=prime_path,
+        )
+    ]
+    window = EchoesGameExportDialog(options, {}, "MyHash", True, games)
 
     # Run
     result = window.get_game_export_params()
@@ -157,4 +189,8 @@ def test_get_game_export_params(skip_qtbot, tmp_path):
         output_path=tmp_path.joinpath("output", "Echoes Randomizer - MyHash.iso"),
         contents_files_path=tmp_path.joinpath("internal_copies", "prime2", "contents"),
         backup_files_path=tmp_path.joinpath("internal_copies", "prime2", "vanilla"),
+        asset_cache_path=tmp_path.joinpath("internal_copies", "prime2", "prime1_models"),
+        prime_path=prime_path,
+        use_prime_models=is_prime_multi and use_external_models,
     )
+