+++ test/generator/filler/test_player_state.py
@@ -1,6 +1,10 @@
+import dataclasses
+from unittest.mock import MagicMock
+
 import pytest
 
 from randovania.game_description import default_database
+from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.games.game import RandovaniaGame
 from randovania.generator.filler import player_state
 from randovania.generator.filler.filler_configuration import FillerConfiguration
@@ -17,6 +21,7 @@ def _default_filler_config() -> FillerConfiguration:
         indices_to_exclude=frozenset(),
         multi_pickup_placement=False,
         logical_resource_action=LayoutLogicalResourceAction.RANDOMLY,
+        first_progression_must_be_local=False,
     )
 
 
@@ -55,3 +60,25 @@ def test_current_state_report(state_for_blank):
         "Intro/Starting Area/Pickup (Weapon)\n"
         "Intro/Starting Area/Pickup (Missile Expansion)"
     )
+
+
+@pytest.mark.parametrize("must_be_local", [False, True])
+@pytest.mark.parametrize("num_assigned_pickups", [0, 1])
+def test_filter_usable_locations(state_for_blank, must_be_local, num_assigned_pickups):
+    state_for_blank.configuration = dataclasses.replace(state_for_blank.configuration,
+                                                        first_progression_must_be_local=must_be_local)
+    state_for_blank.num_assigned_pickups = num_assigned_pickups
+
+    locations_weighted = {
+        (state_for_blank, PickupIndex(0)): 1,
+        (MagicMock(), PickupIndex(0)): 1,
+    }
+
+    # Run
+    filtered = state_for_blank.filter_usable_locations(locations_weighted)
+
+    # Assert
+    if must_be_local and num_assigned_pickups == 0:
+        assert filtered == {(state_for_blank, PickupIndex(0)): 1}
+    else:
+        assert filtered == locations_weighted