+++ test/server/discord/test_database_command.py
@@ -2,21 +2,19 @@ import io
 from pathlib import Path
 from unittest.mock import MagicMock, AsyncMock, ANY, call
 
-import pytest
 from discord_slash import ComponentType
 
 from randovania.games.game import RandovaniaGame
 from randovania.server.discord.database_command import DatabaseCommandCog, SplitWorld
 
 
-async def test_on_ready():
+async def test_add_commands():
     # Setup
     cog = DatabaseCommandCog({"guild": 1234}, MagicMock())
     slash = cog.bot.slash
-    slash.sync_all_commands = AsyncMock()
 
     # Run
-    await cog.on_ready()
+    await cog.add_commands(slash)
 
     # Assert
     slash.add_slash_command.assert_called_once_with(
@@ -31,7 +29,6 @@ async def test_on_ready():
         components=ANY,
         use_callback_name=False,
     )
-    slash.sync_all_commands.assert_awaited_once_with()
 
 
 async def test_database_command():